move things into shared_context
